adk-agent/pyproject.toml
[project]
name = "pokemon-adk-agent"
version = "0.1.0"
description = "Pokemon Agent using Google ADK and Gemini with MCP tools integration"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.1.0",
    "python-dotenv>=1.0.0",
]
